+++ pages/index.js
@@ -1,9 +1,8 @@
-import HeadMeta from '@/components/HeadTag/HeadMeta';
+import HeadMeta from '@/components/headTag/HeadMeta';
 import Education from '@/components/home/Education';
 import Work from '@/components/home/Work';
 import { Box, Flex, Text } from '@chakra-ui/react';
 import { supabase } from 'lib/supabase';
-// import FeaturePost from '@/components/home/FeaturePost';
 
 export default function Home({ workData, educationData }) {
 	return (
