+++ pages/projects/index.js
@@ -1,26 +1,14 @@
-import HeadMeta from '@/components/HeadTag/HeadMeta';
+import HeadMeta from '@/components/headTag/HeadMeta';
+import PageBackground from '@/components/layouts/PageBackground';
 import ProjectSidebar from '@/components/project/ProjectSidebar';
-import { Box, Flex, useColorModeValue } from '@chakra-ui/react';
+import { Flex } from '@chakra-ui/react';
 import getProjects from '../../lib/getProjects';
 
 function Projects() {
-	const color = useColorModeValue('rgba(0,0,0,0.7)', 'rgba(255,255,255,0.5)');
-
 	return (
 		<>
 			<HeadMeta title="Projects" />
-			<Box
-				display={['none', 'none', 'block']}
-				w="70%"
-				pl={5}
-				sx={{
-					backgroundColor: 'rgba(0,0,0,0)',
-					opacity: 0.8,
-					backgroundImage: `radial-gradient(${color} 1px, transparent 1px), radial-gradient(${color} 1px, rgba(0,0,0,0) 1px)`,
-					backgroundSize: '40px 40px',
-					backgroundPosition: '0 0,20px 20px',
-				}}
-			></Box>
+			<PageBackground />
 		</>
 	);
 }
